// src/assets/rawData.d.ts
export declare const rawDataFoodandDrinks: {
    id: string;
    img: string;
    name: string;
    price: number;
    gluten_free: boolean;
    category: string;
}[];
export declare const rawDatacosmetics: {
    id: string;
    img: string;
    name: string;
    price: number;
    gluten_free: boolean;
    category: string;
}[];
export declare const veganPetFood: {
    id: string;
    img: string;
    name: string;
    price: number;
    gluten_free: boolean;
    category: string;
}[];
export declare const data: {
    rawDataFoodandDrinks: {
        id: string;
        img: string;
        name: string;
        price: number;
        gluten_free: boolean;
        category: string;
    }[];
    rawDatacosmetics: {
        id: string;
        img: string;
        name: string;
        price: number;
        gluten_free: boolean;
        category: string;
    }[];
    veganPetFood: {
        id: string;
        img: string;
        name: string;
        price: number;
        gluten_free: boolean;
        category: string;
    }[];
};